pyproject.toml
[tool.poetry]
name = "clangen-wiki-mkdocs"
version = "0.1.0"
description = "Clangen wiki presented via MkDocs."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-material = "^9.5.21"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
